+++ app/[locale]/template/__tests__/page.spec.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import {act, fireEvent, render, screen} from '@/utils/test-utils';
+import { act, fireEvent, render, screen } from '@/utils/test-utils';
 import TemplateListPage from '../page';
-import {axe, toHaveNoViolations} from 'jest-axe';
-import {useTemplatesQuery,} from '@/generated/graphql';
+import { axe, toHaveNoViolations } from 'jest-axe';
+import { useTemplatesQuery, } from '@/generated/graphql';
 
 expect.extend(toHaveNoViolations);
 
@@ -41,7 +41,7 @@ jest.mock('@/components/TemplateListItem', () => {
 
 // Will pass this mock data back when query is made for templates
 const mockTemplateData = {
-  templates: [{
+  myTemplates: [{
     name: 'UCOP',
     description: 'University of California Office of the President',
     modified: '2024-11-20 00:00:00',